+++ src/main/java/roomescape/reservation/application/ReservationFacadeService.java
@@ -13,9 +13,9 @@ import roomescape.member.application.service.MemberQueryService;
 import roomescape.member.domain.Member;
 import roomescape.reservation.application.service.ReservationCommandService;
 import roomescape.reservation.application.service.ReservationQueryService;
-import roomescape.reservation.application.service.WaitingReservationQueryService;
+import roomescape.reservation.waiting.application.service.WaitingReservationQueryService;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.domain.dto.WaitingReservationWithRank;
+import roomescape.reservation.waiting.domain.dto.WaitingReservationWithRank;
 import roomescape.reservation.presentation.dto.MyReservationResponse;
 import roomescape.reservation.presentation.dto.ReservationRequest;
 import roomescape.reservation.presentation.dto.ReservationResponse;
