--- src/main/java/roomescape/reservation/domain/dto/WaitingReservationWithRank.java
@@ -1,6 +1,0 @@
-package roomescape.reservation.domain.dto;
-
-import roomescape.reservation.domain.WaitingReservation;
-
-public record WaitingReservationWithRank(WaitingReservation waitingReservation, long rank) {
-}
