+++ src/main/java/roomescape/reservation/presentation/MemberReservationController.java
@@ -16,7 +16,7 @@ import roomescape.auth.login.presentation.dto.LoginMemberInfo;
 import roomescape.auth.login.presentation.dto.annotation.LoginMember;
 import roomescape.common.exception.handler.dto.ExceptionResponse;
 import roomescape.reservation.application.ReservationFacadeService;
-import roomescape.reservation.application.WaitingReservationFacadeService;
+import roomescape.reservation.waiting.application.WaitingReservationFacadeService;
 import roomescape.reservation.presentation.dto.MyReservationResponse;
 import roomescape.reservation.presentation.dto.ReservationRequest;
 import roomescape.reservation.presentation.dto.ReservationResponse;
