+++ src/main/java/roomescape/reservation/waiting/domain/dto/WaitingReservationWithRank.java
@@ -1,0 +1,6 @@
+package roomescape.reservation.waiting.domain.dto;
+
+import roomescape.reservation.waiting.domain.WaitingReservation;
+
+public record WaitingReservationWithRank(WaitingReservation waitingReservation, long rank) {
+}
