+++ src/test/java/roomescape/reservation/application/ReservationServiceTest.java
@@ -29,14 +29,14 @@ import roomescape.reservation.presentation.dto.MyReservationResponse;
 import roomescape.member.application.service.MemberQueryService;
 import roomescape.reservation.application.service.ReservationCommandService;
 import roomescape.reservation.application.service.ReservationQueryService;
-import roomescape.reservation.application.service.WaitingReservationQueryService;
+import roomescape.reservation.waiting.application.service.WaitingReservationQueryService;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationRepository;
-import roomescape.reservation.domain.WaitingReservationRepository;
+import roomescape.reservation.waiting.domain.WaitingReservationRepository;
 import roomescape.reservation.infrastructure.JpaReservationRepository;
 import roomescape.reservation.infrastructure.JpaReservationRepositoryAdapter;
-import roomescape.reservation.infrastructure.JpaWaitingReservationRepository;
-import roomescape.reservation.infrastructure.JpaWaitingReservationRepositoryAdaptor;
+import roomescape.reservation.waiting.infrastructure.JpaWaitingReservationRepository;
+import roomescape.reservation.waiting.infrastructure.JpaWaitingReservationRepositoryAdaptor;
 import roomescape.reservation.presentation.dto.ReservationRequest;
 import roomescape.reservation.application.ReservationServiceTest.ReservationConfig;
 import roomescape.reservation.time.domain.ReservationTime;
